requirements.txt
langgraph>=0.2.0
langchain-core>=0.2.0
langsmith>=0.1.0
python-dotenv>=1.0.0
pydantic>=2.0.0
google-genai>=0.3.0
azure-search-documents>=11.4.0
pytest>=8.0.0
